refactor: Enable protocol client per namespace for observer
Since transaction processor requires protocol client per namespace split transaction handler into two components:
-- operation handler (creates batch files from batch operations) - called by batch writer
-- operation processor (creates batch operations from batch files) - called by observer

Closes #308
