+++ pkg/mocks/protocol.go
@@ -7,8 +7,13 @@ SPDX-License-Identifier: Apache-2.0
 package mocks
 
 import (
+	"github.com/pkg/errors"
+
 	"github.com/trustbloc/sidetree-core-go/pkg/api/protocol"
 )
+
+// DefaultNS is default namespace used in mocks
+const DefaultNS = "did:sidetree"
 
 // MockProtocolClient mocks protocol for testing purposes.
 type MockProtocolClient struct {
@@ -31,4 +36,29 @@ func NewMockProtocolClient() *MockProtocolClient {
 // Current mocks getting last protocol version
 func (m *MockProtocolClient) Current() protocol.Protocol {
 	return m.Protocol
+}
+
+// NewMockProtocolClientProvider creates new mock protocol client provider
+func NewMockProtocolClientProvider() *MockProtocolClientProvider {
+	m := make(map[string]protocol.Client)
+
+	m[DefaultNS] = NewMockProtocolClient()
+	return &MockProtocolClientProvider{
+		ProtocolClients: m,
+	}
+}
+
+// MockProtocolClientProvider implements mock protocol client provider
+type MockProtocolClientProvider struct {
+	ProtocolClients map[string]protocol.Client
+}
+
+// ForNamespace will return protocol client for that namespace
+func (m *MockProtocolClientProvider) ForNamespace(namespace string) (protocol.Client, error) {
+	pc, ok := m.ProtocolClients[namespace]
+	if !ok {
+		return nil, errors.Errorf("protocol client not found for namespace [%s]", namespace)
+	}
+
+	return pc, nil
 }
