+++ pkg/txnhandler/models/anchor.go
@@ -69,7 +69,7 @@ func getCreateOperations(ops []*batch.Operation) []CreateOperation {
 	var result []CreateOperation
 	for _, op := range ops {
 		if op.Type == batch.OperationTypeCreate {
-			create := CreateOperation{SuffixData: op.EncodedSuffixData}
+			create := CreateOperation{SuffixData: op.EncodedSuffixData, Namespace: op.Namespace}
 
 			result = append(result, create)
 		}
